Revert "render sign in and sign up"
