+++ client/src/routes.js
@@ -11,13 +11,13 @@ import {
 // Admin views
 //import Calendar from "./views/admin/calendar/Calendar.js";
 import Create from "views/admin/create";
-import Journals from "views/admin/journals";
+import Journals from "./views/admin/journals";
 import Dashboard from "views/admin/dashboard";
 import Store from "views/admin/store"
 
 import CancelPage from "views/admin/cancelOrderPage";
 //auth views
-import SignIn from "views/auth/signIn/index.jsx"
+// import SignIn from "./views/admin/auth/SignIn.js";
 // import SignUp from "./views/admin/auth/SignUp.js";
 // import Landing from "./views/admin/auth/Landing.js";
 
@@ -65,13 +65,13 @@ const routes = [
   //     icon: <Icon as={MdOutlineCalendarMonth} width='20px' height='20px' color='inherit' />,
   //     component: Calendar,
   // },
-  {
-      name: "Sign in",
-      layout: "/auth",
-      path: "/sign-in",
-      icon: <Icon as={MdLock} width='20px' height='20px' color='inherit' />,
-      component: SignIn,
-  },
+  // {
+  //     name: "Signin",
+  //     layout: "/auth",
+  //     path: "/sign-in",
+  //     icon: <Icon as={MdLock} width='20px' height='20px' color='inherit' />,
+  //     component: SignIn,
+  // },
   // {
   //     name: "Signup",
   //     layout: "/auth",
