+++ client/src/views/auth/Signin.jsx
@@ -18,15 +18,14 @@ import {
 } from "@chakra-ui/react";
 
 import { HSeparator } from "components/separator/Separator";
-import DefaultAuth from "layouts/auth/Default";
-
 // img
-
+// default auth layout
 import { FcGoogle } from "react-icons/fc";
 import { MdOutlineRemoveRedEye } from "react-icons/md";
 import { RiEyeCloseLine } from "react-icons/ri";
 
-export default function SignUp() {
+function SignIn() {
+    // Chakra color mode
     const textColor = useColorModeValue("navy.700", "white");
     const textColorSecondary = "gray.400";
     const textColorDetails = useColorModeValue("navy.700", "secondaryGray.600");
@@ -44,32 +43,31 @@ export default function SignUp() {
     );
     const [show, setShow] = React.useState(false);
     const handleClick = () => setShow(!show);
-
     return (
-        <DefaultAuth illustrationBackground={illustration} image={illustration}>
-          <Flex
-            maxW={{ base: "100%", md: "max-content" }}
-            w='100%'
-            mx={{ base: "auto", lg: "0px" }}
-            me='auto'
-            h='100%'
-            alignItems='start'
-            justifyContent='center'
-            mb={{ base: "30px", md: "60px" }}
-            px={{ base: "25px", md: "0px" }}
-            mt={{ base: "40px", md: "14vh" }}
-            flexDirection='column'>
-            <Box me='auto'>
-              <Heading color={textColor} fontSize='36px' mb='10px'>
-                Sign Up
-              </Heading>
-              <Text
+      <DefaultAuth illustrationBackground={illustration} image={illustration}>
+        <Flex
+          maxW={{ base: "100%", md: "max-content" }}
+          w='100%'
+          mx={{ base: "auto", lg: "0px" }}
+          me='auto'
+          h='100%'
+          alignItems='start'
+          justifyContent='center'
+          mb={{ base: "30px", md: "60px" }}
+          px={{ base: "25px", md: "0px" }}
+          mt={{ base: "40px", md: "14vh" }}
+          flexDirection='column'>
+             <Box me='auto'>
+          <Heading color={textColor} fontSize='36px' mb='10px'>
+            Sign In
+          </Heading>
+          <Text
             mb='36px'
             ms='4px'
             color={textColorSecondary}
             fontWeight='400'
             fontSize='md'>
-            Enter your email and password to sign up!
+            Enter your email and password to sign in!
           </Text>
         </Box>
         <Flex
@@ -82,7 +80,7 @@ export default function SignUp() {
           mx={{ base: "auto", lg: "unset" }}
           me='auto'
           mb={{ base: "20px", md: "auto" }}>
-             <Button
+          <Button
             fontSize='sm'
             me='0px'
             mb='26px'
@@ -90,13 +88,13 @@ export default function SignUp() {
             h='50px'
             borderRadius='16px'
             bg={googleBg}
-          color={googleText}
-          fontWeight='500'
-          _hover={googleHover}
-          _active={googleActive}
-          _focus={googleActive}>
-          <Icon as={FcGoogle} w='20px' h='20px' me='10px' />
-            Sign up with Google
+            color={googleText}
+            fontWeight='500'
+            _hover={googleHover}
+            _active={googleActive}
+            _focus={googleActive}>
+            <Icon as={FcGoogle} w='20px' h='20px' me='10px' />
+            Sign in with Google
           </Button>
           <Flex align='center' mb='25px'>
             <HSeparator />
@@ -123,38 +121,10 @@ export default function SignUp() {
               type='email'
                 placeholder='Enter your email'
                 mb='24px'
-                fontWeight='500'
-                size='lg'
-              />
-              <FormLabel
-                ms='4px'
-                fontSize='sm'
-                fontWeight='500'
-                color={textColor}
-                display='flex'>
-                Password<Text color={brandStars}>*</Text>
-              </FormLabel>
-              <InputGroup size='md'>
-              <Input
-                isRequired={true}
-                fontSize='sm'
-                placeholder='Min. 8 characters'
-                mb='12px'
-                size='lg'
-                type={show ? "text" : "password"}
-                variant='auth'
-              />
-              <InputRightElement display='flex' alignItems='center' mt='4px'>
-                <Icon
-                  color={textColorSecondary}
-                  _hover={{ cursor: "pointer" }}
-                  as={show ? RiEyeCloseLine : MdOutlineRemoveRedEye}
-                  onClick={handleClick}
-                />
-              </InputRightElement>
-            </InputGroup>
-
-            <FormLabel
+              fontWeight='500'
+              size='lg'
+            />
+             <FormLabel
               ms='4px'
               fontSize='sm'
               fontWeight='500'
@@ -167,33 +137,6 @@ export default function SignUp() {
                 isRequired={true}
                 fontSize='sm'
                 placeholder='Min. 8 characters'
-                mb='12px'
-                size='lg'
-                type={show ? "text" : "password"}
-                variant='auth'
-              />
-              <InputRightElement display='flex' alignItems='center' mt='4px'>
-                <Icon
-                  color={textColorSecondary}
-                  _hover={{ cursor: "pointer" }}
-                  as={show ? RiEyeCloseLine : MdOutlineRemoveRedEye}
-                  onClick={handleClick}
-                />
-              </InputRightElement>
-            </InputGroup>
-            <FormLabel
-              ms='4px'
-              fontSize='sm'
-              fontWeight='500'
-              color={textColor}
-              display='flex'>
-              Confirm Password<Text color={brandStars}>*</Text>
-            </FormLabel>
-            <InputGroup size='md'>
-              <Input
-                isRequired={true}
-                fontSize='sm'
-                placeholder='Confirm your password'
                 mb='24px'
                 size='lg'
                 type={show ? "text" : "password"}
@@ -224,7 +167,7 @@ export default function SignUp() {
                   Keep me logged in
                 </FormLabel>
               </FormControl>
-            </Flex>
+              </Flex>
             <Button
               fontSize='sm'
               variant='brand'
@@ -232,7 +175,7 @@ export default function SignUp() {
               w='100%'
               h='50'
               mb='24px'>
-              Sign Up
+              Sign In
             </Button>
           </FormControl>
           <Flex
@@ -242,14 +185,14 @@ export default function SignUp() {
             maxW='100%'
             mt='0px'>
             <Text color={textColorDetails} fontWeight='400' fontSize='14px'>
-              Already have an account?
-              <NavLink to='/auth/sign-in'>
+              Not registered yet?
+              <NavLink to='/auth/sign-up'>
                 <Text
                   color={textColorBrand}
                   as='span'
                   ms='5px'
                   fontWeight='500'>
-                  Sign In
+                  Create an Account
                 </Text>
               </NavLink>
             </Text>
@@ -259,3 +202,5 @@ export default function SignUp() {
     </DefaultAuth>
   );
 }
+
+export default SignIn;
