+++ client/src/views/auth/Signup.jsx
@@ -1,0 +1,263 @@
+import React from "react";
+import { NavLink } from "react-router-dom";
+// Chakra imports
+import {
+  Box,
+  Button,
+  Checkbox,
+  Flex,
+  FormControl,
+  FormLabel,
+  Heading,
+  Icon,
+  Input,
+  InputGroup,
+  InputRightElement,
+  Text,
+  useColorModeValue,
+} from "@chakra-ui/react";
+
+import { HSeparator } from "components/separator/Separator";
+import DefaultAuth from "layouts/auth/Default";
+
+// img
+
+import { FcGoogle } from "react-icons/fc";
+import { MdOutlineRemoveRedEye } from "react-icons/md";
+import { RiEyeCloseLine } from "react-icons/ri";
+
+function SignUp() {
+    const textColor = useColorModeValue("navy.700", "white");
+    const textColorSecondary = "gray.400";
+    const textColorDetails = useColorModeValue("navy.700", "secondaryGray.600");
+    const textColorBrand = useColorModeValue("brand.500", "white");
+    const brandStars = useColorModeValue("brand.500", "brand.400");
+    const googleBg = useColorModeValue("secondaryGray.300", "whiteAlpha.200");
+    const googleText = useColorModeValue("navy.700", "white");
+    const googleHover = useColorModeValue(
+      { bg: "gray.200" },
+      { bg: "whiteAlpha.300" }
+    );
+    const googleActive = useColorModeValue(
+      { bg: "secondaryGray.300" },
+      { bg: "whiteAlpha.200" }
+    );
+    const [show, setShow] = React.useState(false);
+    const handleClick = () => setShow(!show);
+
+    return (
+        <DefaultAuth illustrationBackground={illustration} image={illustration}>
+          <Flex
+            maxW={{ base: "100%", md: "max-content" }}
+            w='100%'
+            mx={{ base: "auto", lg: "0px" }}
+            me='auto'
+            h='100%'
+            alignItems='start'
+            justifyContent='center'
+            mb={{ base: "30px", md: "60px" }}
+            px={{ base: "25px", md: "0px" }}
+            mt={{ base: "40px", md: "14vh" }}
+            flexDirection='column'>
+            <Box me='auto'>
+              <Heading color={textColor} fontSize='36px' mb='10px'>
+                Sign Up
+              </Heading>
+              <Text
+            mb='36px'
+            ms='4px'
+            color={textColorSecondary}
+            fontWeight='400'
+            fontSize='md'>
+            Enter your email and password to sign up!
+          </Text>
+        </Box>
+        <Flex
+          zIndex='2'
+          direction='column'
+          w={{ base: "100%", md: "420px" }}
+          maxW='100%'
+          background='transparent'
+          borderRadius='15px'
+          mx={{ base: "auto", lg: "unset" }}
+          me='auto'
+          mb={{ base: "20px", md: "auto" }}>
+             <Button
+            fontSize='sm'
+            me='0px'
+            mb='26px'
+            py='15px'
+            h='50px'
+            borderRadius='16px'
+            bg={googleBg}
+          color={googleText}
+          fontWeight='500'
+          _hover={googleHover}
+          _active={googleActive}
+          _focus={googleActive}>
+          <Icon as={FcGoogle} w='20px' h='20px' me='10px' />
+            Sign up with Google
+          </Button>
+          <Flex align='center' mb='25px'>
+            <HSeparator />
+            <Text color='gray.400' mx='14px'>
+              or
+            </Text>
+            <HSeparator />
+          </Flex>
+          <FormControl>
+            <FormLabel
+              display='flex'
+              ms='4px'
+              fontSize='sm'
+              fontWeight='500'
+              color={textColor}
+              mb='8px'>
+              Email<Text color={brandStars}>*</Text>
+            </FormLabel>
+            <Input
+              isRequired={true}
+              variant='auth'
+              fontSize='sm'
+              ms={{ base: "0px", md: "0px" }}
+              type='email'
+                placeholder='Enter your email'
+                mb='24px'
+                fontWeight='500'
+                size='lg'
+              />
+              <FormLabel
+                ms='4px'
+                fontSize='sm'
+                fontWeight='500'
+                color={textColor}
+                display='flex'>
+                Password<Text color={brandStars}>*</Text>
+              </FormLabel>
+              <InputGroup size='md'>
+              <Input
+                isRequired={true}
+                fontSize='sm'
+                placeholder='Min. 8 characters'
+                mb='12px'
+                size='lg'
+                type={show ? "text" : "password"}
+                variant='auth'
+              />
+              <InputRightElement display='flex' alignItems='center' mt='4px'>
+                <Icon
+                  color={textColorSecondary}
+                  _hover={{ cursor: "pointer" }}
+                  as={show ? RiEyeCloseLine : MdOutlineRemoveRedEye}
+                  onClick={handleClick}
+                />
+              </InputRightElement>
+            </InputGroup>
+
+            <FormLabel
+              ms='4px'
+              fontSize='sm'
+              fontWeight='500'
+              color={textColor}
+              display='flex'>
+              Password<Text color={brandStars}>*</Text>
+            </FormLabel>
+            <InputGroup size='md'>
+              <Input
+                isRequired={true}
+                fontSize='sm'
+                placeholder='Min. 8 characters'
+                mb='12px'
+                size='lg'
+                type={show ? "text" : "password"}
+                variant='auth'
+              />
+              <InputRightElement display='flex' alignItems='center' mt='4px'>
+                <Icon
+                  color={textColorSecondary}
+                  _hover={{ cursor: "pointer" }}
+                  as={show ? RiEyeCloseLine : MdOutlineRemoveRedEye}
+                  onClick={handleClick}
+                />
+              </InputRightElement>
+            </InputGroup>
+            <FormLabel
+              ms='4px'
+              fontSize='sm'
+              fontWeight='500'
+              color={textColor}
+              display='flex'>
+              Confirm Password<Text color={brandStars}>*</Text>
+            </FormLabel>
+            <InputGroup size='md'>
+              <Input
+                isRequired={true}
+                fontSize='sm'
+                placeholder='Confirm your password'
+                mb='24px'
+                size='lg'
+                type={show ? "text" : "password"}
+                variant='auth'
+              />
+              <InputRightElement display='flex' alignItems='center' mt='4px'>
+                <Icon
+                  color={textColorSecondary}
+                  _hover={{ cursor: "pointer" }}
+                  as={show ? RiEyeCloseLine : MdOutlineRemoveRedEye}
+                  onClick={handleClick}
+                />
+              </InputRightElement>
+            </InputGroup>
+            <Flex justifyContent='space-between' align='center' mb='24px'>
+              <FormControl display='flex' alignItems='center'>
+                <Checkbox
+                  id='remember-login'
+                  colorScheme='brandScheme'
+                  me='10px'
+                />
+                <FormLabel
+                  htmlFor='remember-login'
+                  mb='0'
+                  fontWeight='normal'
+                  color={textColor}
+                  fontSize='sm'>
+                  Keep me logged in
+                </FormLabel>
+              </FormControl>
+            </Flex>
+            <Button
+              fontSize='sm'
+              variant='brand'
+              fontWeight='500'
+              w='100%'
+              h='50'
+              mb='24px'>
+              Sign Up
+            </Button>
+          </FormControl>
+          <Flex
+            flexDirection='column'
+            justifyContent='center'
+            alignItems='start'
+            maxW='100%'
+            mt='0px'>
+            <Text color={textColorDetails} fontWeight='400' fontSize='14px'>
+              Already have an account?
+              <NavLink to='/auth/sign-in'>
+                <Text
+                  color={textColorBrand}
+                  as='span'
+                  ms='5px'
+                  fontWeight='500'>
+                  Sign In
+                </Text>
+              </NavLink>
+            </Text>
+          </Flex>
+        </Flex>
+      </Flex>
+    </DefaultAuth>
+  );
+}
+
+export default SignUp;
