+++ client/src/views/auth/index.jsx
@@ -1,0 +1,18 @@
+import React from "react";
+import { NavLink } from "react-router-dom";
+// Chakra imports
+import {
+  Box,
+  Button,
+  Checkbox,
+  Flex,
+  FormControl,
+  FormLabel,
+  Heading,
+  Icon,
+  Input,
+  InputGroup,
+  InputRightElement,
+  Text,
+  useColorModeValue,
+} from "@chakra-ui/react";
